Add a "last block notifier queue"
Due to the buffer, the "read" cursor is different than the "play" cursor.
The Player needs to know which audio block is the final block to reset the "play" cursor.
Another queue is used, where booleans are passed from the Reader indicating whether it is the final block before looping back.